refactor(question): extract DuplicateService from QuestionService
Move the duplicate-detection methods (checkDuplicateQuestion,
manualCheckDuplicate) into a new DuplicateService via facade + delegation.
The GDB -> pending-queue -> LLM pipeline (runDuplicateCheckPipeline), which
is shared with QuestionService's addQuestion background flow, is extracted to
a stateless helper (services/helpers/duplicatePipeline.ts) taking aiService,
so both call sites reuse one implementation. Dropped the now-unused
checkDuplicateQuestionHelper / checkConceptDuplicate / isQuestionMatchForPaeExpert
imports from QuestionService.
