+++ backend/src/modules/core/container.ts
@@ -18,6 +18,7 @@ import { QuestionReportService } from '../question/services/QuestionReportServic
 import { PaeValidationService } from '../question/services/PaeValidationService.js';
 import { FeedbackService } from '../question/services/FeedbackService.js';
 import { QuestionAiService } from '../question/services/QuestionAiService.js';
+import { DuplicateService } from '../question/services/DuplicateService.js';
 import { ContextController } from '../context/controllers/ContextController.js';
 import { ContextService } from '../context/services/ContextService.js';
 import { PerformanceController } from '../performance/controllers/PerformanceController.js';
@@ -60,6 +61,10 @@ export const coreContainerModule = new ContainerModule(options => {
   options
     .bind(GLOBAL_TYPES.QuestionAiService)
     .to(QuestionAiService)
+    .inSingletonScope();
+  options
+    .bind(GLOBAL_TYPES.DuplicateService)
+    .to(DuplicateService)
     .inSingletonScope();
   options.bind(CORE_TYPES.AnswerService).to(AnswerService).inSingletonScope();
   options.bind(CORE_TYPES.ContextService).to(ContextService).inSingletonScope();
